+++ robot_descriptions/_descriptions.py
@@ -121,6 +121,7 @@ DESCRIPTIONS: Dict[str, Description] = {
     "shadow_hand_mj_description": Description(Format.MJCF),
     "simple_humanoid_description": Description(Format.URDF),
     "solo_description": Description(Format.URDF),
+    "spryped_description": Description(Format.URDF),
     "talos_description": Description(Format.URDF),
     "tiago_description": Description(Format.URDF),
     "upkie_description": Description(Format.URDF),