+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## Unreleased
+
+### Added
+
+- Description: Spryped
+
 ## [1.0.0] - 2022/10/31
 
 ### Added